Temporarily disabled continue button in main menu
The continue button did not have a similar icon to the other menu items and as a result, it messed up the formatting of the menu when it appeared. I've disabled the button and commented out that code until/unless our game is actually long enough with save states that we need a continue button.