Add area interaction progress: toggle labels
Note: I'm not initializing the player variable with @onready anymore because our interaction manager is a singleton. Using @onready when starting from the menu later resulted in it being null. Is there a standard setup for singletons accessing variables we expect to be initialized during certain scenes? Otherwise my intuition is to take it out of being a singleton and add it to the Office overworld scene.